Improved rotating statuses
Rotating statuses can now be toggled between sequential and random
orders. An initial set of statuses is loaded from a file on disk, so
that large numbers of statuses do not need to be re-entered each time.
At present statuses created within the bot do not save to this list.