Partial restructuring and inventory system
Lots of adjustments which will, hopefully, make the code easier to
manage going forwards. Splitting stuff between lots of different files
to minimize the amount of stuff in any one place and to group them to
gether in a logical manner, that sort of thing.

Also redid the purchase system for drinks. Now they're added to the
user's inventory when purchased, and can be given as gifts or drunk.